Added functionality for more GUI stuff and renamed the block class to blockgraphics
