+++ TetrisProject/TetrisProject.csproj
@@ -45,7 +45,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Block.cs" />
+    <Compile Include="BlockGraphics.cs" />
     <Compile Include="Board.cs" />
     <Compile Include="GameConstants.cs" />
     <Compile Include="VisualConstants.cs" />
